new update in navbar
