+++ src/shared/index.js
@@ -1,1 +1,7 @@
-export { MainBtn, SecondaryBtn } from "./Btns";
+export { AccentBtn, SecondaryBtn, HamburgerMobileBtn } from "./Btns";
+
+// Images
+export { default as logo_black } from "./assets/logo_black.svg";
+export { default as logo_white } from "./assets/logo_white.svg";
+export { default as name_logo_black } from "./assets/name_logo_black.svg";
+export { default as name_logo_white } from "./assets/name_logo_white.svg";
