+++ src/widgets/home/Navbar/BttmNavDesktop.jsx
@@ -1,0 +1,143 @@
+import React from "react";
+import styled from "styled-components";
+import { Link } from "react-router-dom";
+
+// Components
+import { AccentBtn } from "../../../shared";
+import { name_logo_black } from "../../../shared";
+
+const BottomNavDesktop = () => {
+  return (
+    <StyledWrapper>
+      <nav className="bottom">
+        <div>
+          <div className="logo">
+            <Link to="/">
+              <svg
+                width="120"
+                height="69"
+                viewBox="0 0 365 69"
+                fill="none"
+                xmlns="http://www.w3.org/2000/svg"
+              >
+                <path
+                  fill-rule="evenodd"
+                  clip-rule="evenodd"
+                  d="M119 33.5V67H124C127.778 67 129 66.633 129 65.5C129 63.573 129.302 63.603 134 66C147.259 72.764 163 61.826 163 45.848C163 30.127 145.426 18.755 133.25 26.597L131 28.046V14.023V0H125H119V33.5ZM218.71 0.623001C218.32 1.014 218 16.108 218 34.167V67H224.513H231.027L230.763 33.75L230.5 0.5L224.96 0.206001C221.913 0.0450013 219.101 0.232001 218.71 0.623001ZM270 14.023V28.046L267.75 26.597C258.663 20.744 243.642 26.516 239.459 37.467C236.955 44.025 237.778 53.276 241.341 58.615C246.956 67.03 258.961 70.449 267.265 65.998C270.365 64.336 271 64.251 271 65.498C271 66.652 272.275 67 276.5 67H282V33.5V0H276H270V14.023ZM83.5 25.061C81.85 25.545 78.853 27.304 76.84 28.97C64.611 39.086 67.183 59.635 81.491 66.135C87.079 68.674 98.088 68.597 103.857 65.979C106.802 64.643 108.037 63.538 107.666 62.572C107.364 61.785 106.816 59.535 106.447 57.571C106.079 55.607 105.601 54 105.385 54C105.169 54 103.314 54.9 101.261 56C96.696 58.447 88.407 58.674 85.217 56.439C83.991 55.581 82.362 53.668 81.597 52.189L80.205 49.5L96.102 49.838L112 50.176L111.978 45.338C111.911 30.662 98.048 20.791 83.5 25.061ZM180.26 26.459C171.706 30.442 166.843 40.19 168.504 50.025C169.449 55.614 174.682 62.94 179.5 65.417C185.473 68.487 196.367 68.816 202.483 66.11C206.788 64.206 207.16 63.775 206.562 61.391C206.2 59.948 205.648 57.489 205.335 55.926L204.767 53.086L201.151 55.543C197.172 58.247 189.629 58.9 185.684 56.883C183.25 55.638 180.057 52.039 180.021 50.5C180.007 49.875 185.9 49.5 195.749 49.5H211.5L211.214 44C210.81 36.225 207.071 30.425 200.12 26.794C193.307 23.235 187.4 23.135 180.26 26.459ZM303.01 25.316C297.973 26.92 291.368 33.923 289.952 39.161C287.248 49.168 290.887 59.25 299.185 64.741C304.963 68.565 316.471 69.212 323.489 66.108C326.738 64.67 328.047 63.565 327.669 62.579C327.365 61.788 326.816 59.535 326.447 57.571C326.079 55.607 325.543 54 325.257 54C324.97 54 323.195 54.952 321.312 56.116C318.52 57.842 316.763 58.151 311.793 57.791C306.69 57.422 305.316 56.906 303.349 54.62C299.098 49.678 299.661 49.472 316.499 49.82L332 50.141V45.135C332 38.798 330.7 35.37 326.689 31.128C320.612 24.702 311.76 22.529 303.01 25.316ZM0 25.855C0 26.855 18.909 67.159 19.872 68.212C20.23 68.603 23.499 63.751 27.137 57.428L33.75 45.933L39.36 55.754C42.445 61.156 45.539 66.331 46.235 67.253C47.327 68.702 48.73 66.328 56.5 49.885C61.45 39.41 66.064 29.526 66.753 27.92L68.006 25H61.963H55.919L50.793 35.544L45.667 46.087L40.177 35.544C37.158 29.745 34.41 25 34.072 25C33.733 25 31.08 29.777 28.177 35.615C25.273 41.454 22.653 45.987 22.355 45.689C22.057 45.391 19.934 41.176 17.637 36.323C15.341 31.47 13.106 26.938 12.672 26.25C11.778 24.835 0 24.468 0 25.855ZM339 46V67H344.918H350.837L351.168 54.563C351.494 42.341 351.551 42.072 354.458 39.041C356.322 37.097 358.817 35.733 361.208 35.351C363.293 35.017 364.997 34.239 364.994 33.622C364.978 30.405 362.738 25 361.421 25C358.605 25 353.257 27.795 351.631 30.117L350 32.445V28.723V25H344.5H339V46ZM98.327 37.25L101.5 40.5L95.713 40.801C92.531 40.967 87.581 40.967 84.713 40.801L79.5 40.5L82.673 37.25C85.355 34.503 86.567 34 90.5 34C94.433 34 95.645 34.503 98.327 37.25ZM197.195 36.687C198.738 38.165 200 39.74 200 40.187C200 40.634 195.5 41 190 41C184.5 41 180 40.611 180 40.136C180 39.661 181.186 38.086 182.636 36.636C186.394 32.879 193.245 32.902 197.195 36.687ZM316.77 35.551C317.988 36.404 319.454 37.979 320.027 39.051C321.029 40.923 320.672 41 311.035 41C304.721 41 301 40.612 301 39.955C301 39.38 302.105 37.805 303.455 36.455C306.37 33.539 313.232 33.073 316.77 35.551ZM146.783 36.561C152.609 40.642 153.349 48.939 148.35 54.156C143.526 59.192 134.373 57.683 131.117 51.316C127.492 44.228 132.9 35 140.677 35C142.81 35 145.558 35.702 146.783 36.561ZM265.783 36.561C271.609 40.642 272.349 48.939 267.35 54.156C262.526 59.192 253.373 57.683 250.117 51.316C246.492 44.228 251.9 35 259.677 35C261.81 35 264.558 35.702 265.783 36.561Z"
+                  fill="black"
+                />
+              </svg>
+            </Link>
+          </div>
+          <div className="navigation">
+            <div className="menubar">
+              <span>Products</span>
+              <div className="menu"></div>
+            </div>
+            <div className="navigation">
+              <Link>Pricing</Link>
+            </div>
+            <div className="navigation">
+              <Link>Solutions</Link>
+            </div>
+            <div className="navigation">
+              <Link>Resources</Link>
+            </div>
+          </div>
+          <div className="btn_group">
+            <AccentBtn content="Сontact us" link="/" />
+          </div>
+        </div>
+      </nav>
+    </StyledWrapper>
+  );
+};
+
+const StyledWrapper = styled.div`
+  .bottom {
+    position: relative;
+    margin: 0 auto;
+    max-width: 1120px;
+
+    div {
+      display: flex;
+      justify-content: space-between;
+      align-items: center;
+      gap: 20px;
+
+      .navigation {
+        -moz-box-align: stretch;
+        align-items: stretch;
+        align-self: stretch;
+        display: flex;
+        -moz-box-pack: center;
+        justify-content: center;
+        margin: 0px auto;
+
+        a {
+          cursor: pointer;
+        }
+        span {
+          cursor: default;
+        }
+
+        a,
+        span {
+          -moz-box-align: center;
+          align-items: center;
+          font-size: 16px;
+          border-radius: 0px;
+          font-stretch: expanded;
+          font-weight: 500;
+          display: flex;
+          line-height: calc(1.25em + 6px);
+          flex-shrink: 0;
+          padding: 16px;
+          position: relative;
+          color: var(--black);
+        }
+
+        a:hover:after,
+        span:hover:after {
+          width: 100%;
+        }
+
+        a:after,
+        span:after {
+          background: rgb(22, 20, 12);
+          bottom: 0px;
+          content: "";
+          display: block;
+          height: 3px;
+          left: 0px;
+          position: absolute;
+          transition: width 0.25s cubic-bezier(0.7, 0, 0.3, 1);
+          width: 0%;
+        }
+
+        .menubar {
+          .menu {
+            max-width: 1120px;
+            min-width: 320px;
+            position: absolute;
+            top: calc(100% + 1px);
+            transition: visibility 0s 250ms, opacity 0s 250ms;
+            width: 100%;
+            z-index: 20;
+            box-shadow: rgba(0, 0, 0, 0.15) 0px 16px 24px;
+            background: var(--white);
+            left: 0px;
+            display: none;
+            height: 200px;
+          }
+        }
+
+        .menubar:hover .menu {
+          display: block;
+        }
+      }
+    }
+  }
+`;
+
+export default BottomNavDesktop;
