+++ src/widgets/home/Navbar/index.js
@@ -1,2 +1,6 @@
+// Desktop
 export { default as TopNavDesktop } from "./TopNavDesktop";
-export { default as BottomNavDesktop } from "./BttmNavMobile";
+export { default as BottomNavDesktop } from "./BttmNavDesktop";
+
+// Mobile
+export { default as TopNavMobile } from "./TopNavMobile";
